[WIP] many-to-many connections support
This patch is an attempt at adding many-to-many connection support
without overhauling the library and minimal breaking changes to the
user.

The core of the idea is to replace Graph::connections from
SecondaryMap<InputId, OutputId> to [..]<InputId, HashSet<OutputId>> and
roll from that.

Ports are parametrized by an argument specifying the maximum non-zero
number of connections (not yet enforced), and if that number is greater
than 1 a so called wide-port is drawn -- it's characterized by being
longer than a typical port and connections made to it are spread out and
can be disconnected individually.

Known limitations:
* Again, max connections per node is not enforced.
* HashSet isn't an adequate data structure for storing the connected
  outputs -- we have no control over ordering, let alone the end user.
* ...?